+++ zlib-rs/src/deflate/algorithm/rle.rs
@@ -17,9 +17,9 @@ pub fn deflate_rle(stream: &mut DeflateStream, flush: DeflateFlush) -> BlockStat
         // at the end of the input file. We need STD_MAX_MATCH bytes
         // for the next match, plus WANT_MIN_MATCH bytes to insert the
         // string following the next match.
-        if stream.state.lookahead < MIN_LOOKAHEAD {
+        if stream.state.lookahead < MIN_LOOKAHEAD as u32 {
             fill_window(stream);
-            if stream.state.lookahead < MIN_LOOKAHEAD && flush == DeflateFlush::NoFlush {
+            if stream.state.lookahead < MIN_LOOKAHEAD as u32 && flush == DeflateFlush::NoFlush {
                 return BlockState::NeedMore;
             }
             if stream.state.lookahead == 0 {
@@ -29,19 +29,19 @@ pub fn deflate_rle(stream: &mut DeflateStream, flush: DeflateFlush) -> BlockStat
 
         /* See how many times the previous byte repeats */
         let state = &mut stream.state;
-        if state.lookahead >= STD_MIN_MATCH && state.strstart > 0 {
-            let scan = &state.window.filled()[state.strstart - 1..][..3 + 256];
+        if state.lookahead >= STD_MIN_MATCH as u32 && state.strstart > 0 {
+            let scan = &state.window.filled()[state.strstart as usize - 1..][..3 + 256];
 
             {
                 if scan[0] == scan[1] && scan[1] == scan[2] {
                     match_len = compare256_rle_slice(scan[0], &scan[3..]) + 2;
-                    match_len = Ord::min(match_len, state.lookahead);
+                    match_len = Ord::min(match_len, state.lookahead as usize);
                     match_len = Ord::min(match_len, STD_MAX_MATCH);
                 }
             }
 
             assert!(
-                state.strstart - 1 + match_len <= state.window_size - 1,
+                state.strstart as usize - 1 + match_len <= state.window_size as usize - 1,
                 "wild scan"
             );
         }
@@ -52,12 +52,12 @@ pub fn deflate_rle(stream: &mut DeflateStream, flush: DeflateFlush) -> BlockStat
 
             bflush = state.tally_dist(1, match_len - STD_MIN_MATCH);
 
-            state.lookahead -= match_len;
-            state.strstart += match_len;
+            state.lookahead -= match_len as u32;
+            state.strstart += match_len as u32;
             match_len = 0;
         } else {
             /* No match, output a literal byte */
-            let lc = state.window.filled()[state.strstart];
+            let lc = state.window.filled()[state.strstart as usize];
             bflush = state.tally_lit(lc);
             state.lookahead -= 1;
             state.strstart += 1;
